implemented tabs

+++ src/com/vortexwolf/dvach/common/utils/IoUtils.java
@@ -2,14 +2,16 @@ package com.vortexwolf.dvach.common.utils;
 
 import java.io.File;
 
+import com.vortexwolf.dvach.settings.ApplicationSettings;
+
 import android.app.Application;
-import android.content.pm.PackageManager;
-import android.content.pm.PackageStats;
 import android.os.Environment;
 
 public class IoUtils {
 
 	public static long dirSize(File dir) {
+		if(dir == null) return 0;
+		
 	    long result = 0;
 	    File[] fileList = dir.listFiles();
 
@@ -25,43 +27,21 @@ public class IoUtils {
 	    return result; // return the file size
 	}
 	
-	public static boolean deleteDirectory(File path) {
-	    if(path.exists()) {
-	      File[] files = path.listFiles();
-	      if (files == null) {
-	          return true;
-	      }
-	      for(int i=0; i<files.length; i++) {
-	         if(files[i].isDirectory()) {
-	           deleteDirectory(files[i]);
-	         }
-	         else {
-	           files[i].delete();
-	         }
-	      }
-	    }
-	    return(path.delete());
-	}
-	
-	public static File tryGetExternalCachePath(Application app){
-		// Check if the external storage is writeable
-		if (Environment.MEDIA_MOUNTED.equals(Environment.getExternalStorageState()))
-		{
-			// Retrieve the base path for the application in the external storage
-			File externalStorageDir = Environment.getExternalStorageDirectory();
-			// {SD_PATH}/Android/data/com.vortexwolf.dvach/cache
-			File extStorageAppCachePath = new File(externalStorageDir,
-					"Android" + File.separator + "data" + File.separator + app.getPackageName() + File.separator + "cache");
-
-			if (!extStorageAppCachePath.exists())
-			{
-				extStorageAppCachePath.mkdirs();
+	public static void deleteDirectory(File path) {
+		if (path != null && path.exists()) {
+			File[] files = path.listFiles();
+			if (files == null) {
+				return;
 			}
-			
-			return extStorageAppCachePath;
-		}
-		
-		return null;
-	}
+			for (int i = 0; i < files.length; i++) {
+				if (files[i].isDirectory()) {
+					deleteDirectory(files[i]);
+				} else {
+					files[i].delete();
+				}
+			}
 
+			path.delete();
+		}
+	}
 }
